MTSDK-221 Refactor MessagingClientImplTest.kt (#208)
* MTSDK-221 Refactor MessagingClientImplTest.kt

- Break down MessagingClientImpl tests to separate files organized by feature.
